Move StubContainer to graph
