+++ tests/unit/source_code/test_files.py
@@ -6,8 +6,8 @@ import pytest
 from databricks.sdk.service.workspace import Language
 
 from databricks.labs.ucx.source_code.base import CurrentSessionState
-from databricks.labs.ucx.source_code.graph import Dependency, DependencyGraph, DependencyProblem
-from databricks.labs.ucx.source_code.files import FileLoader, LocalFile, StubContainer
+from databricks.labs.ucx.source_code.graph import Dependency, DependencyGraph, DependencyProblem, StubContainer
+from databricks.labs.ucx.source_code.files import FileLoader, LocalFile
 from databricks.labs.ucx.source_code.path_lookup import PathLookup
